refactor: relocate health service
